+++ RevitAddin.Icon.Example/Revit/App.cs
@@ -2,6 +2,7 @@ using Autodesk.Revit.DB;
 using Autodesk.Revit.UI;
 using ricaun.Revit.UI;
 using System.IO;
+using System.Linq;
 
 namespace RevitAddin.Icon.Example.Revit
 {
@@ -9,23 +10,43 @@ namespace RevitAddin.Icon.Example.Revit
     public class App : IExternalApplication
     {
         private static RibbonPanel ribbonPanel;
-        private static RibbonImageThemeSelector ribbonImageThemeSelector;
         public Result OnStartup(UIControlledApplication application)
         {
-            ribbonImageThemeSelector = new RibbonImageThemeSelector(application);
-
-            var icons = new[] { "Grey", "Red", "Yellow", "Green", "Cyan", "Blue", "Purple", "Pink", "Brown" };
+            var icons = new[] { "Grey", "Red", "Orange", "Yellow", "Green", "Cyan", "Blue", "Purple", "Pink", "Brown" };
 
             ribbonPanel = application.CreatePanel("RevitAddin.Icon.Example");
+
             foreach (var icon in icons)
             {
                 ribbonPanel.CreatePushButton<Commands.Command>(icon)
-                    .SetLargeImage($"Resources/Images/Cube-{icon}.ico");
+                    .SetLargeImage($"Resources/Images/Cube-{icon}-Light.tiff");
             }
 
-            ribbonImageThemeSelector.AddRibbonItem(ribbonPanel.GetRibbonItems());
+            ribbonPanel.AddSlideOut();
 
-            ribbonImageThemeSelector.UpdateImages();
+            foreach (var icon in icons)
+            {
+                ribbonPanel.RowLargeStackedItems(
+                    ribbonPanel.CreatePushButton<Commands.Command>(icon)
+                        .SetLargeImage($"Resources/Images/Cube-{icon}-Light.tiff")
+                        .SetShowText(),
+                    ribbonPanel.CreatePushButton<Commands.Command>(icon)
+                        .SetLargeImage($"Resources/Images/Box-{icon}-Light.tiff")
+                        .SetShowText()
+                );
+            }
+
+            foreach (var icon in icons)
+            {
+                ribbonPanel.RowStackedItems(
+                    ribbonPanel.CreatePushButton<Commands.Command>(icon)
+                        .SetLargeImage($"Resources/Images/Cube-{icon}-Light.tiff")
+                        .SetShowText(),
+                    ribbonPanel.CreatePushButton<Commands.Command>(icon)
+                        .SetLargeImage($"Resources/Images/Box-{icon}-Light.tiff")
+                        .SetShowText()
+                );
+            }
 
             return Result.Succeeded;
         }
@@ -33,7 +54,6 @@ namespace RevitAddin.Icon.Example.Revit
         public Result OnShutdown(UIControlledApplication application)
         {
             ribbonPanel?.Remove();
-            ribbonImageThemeSelector?.Dispose();
             return Result.Succeeded;
         }
     }