+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] / 2024-11-14
+- Update `ricaun.Revit.UI` to `0.7.1`
+- Add missing `Orange` color
+- Obsolete `RibbonImageThemeSelector` to use `ricaun.Revit.UI`.
+- Rename Images to have `Light` in the name to enable the theme change to `Dark`.
+- Add `Box` image files as resources.
+- Use `tiff` files to enable high-quality icon.
+
 ## [1.0.1] / 2024-01-11
 - Add Revit version 2025 configuration
 
@@ -11,5 +19,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.1.0]: ../../compare/1.0.1...1.1.0
 [1.0.1]: ../../compare/1.0.0...1.0.1
 [1.0.0]: ../../compare/1.0.0